% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{ATS}
\alias{ATS}
\title{Anova-Type-statistic}
\usage{
ATS(N, vVarData, C, HatCov, Xi = 0)

ATS(N, vVarData, C, HatCov, Xi = 0)
}
\description{
Anova-Type-statistic

Anova-Type-statistic
}
