% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{ATShtilde}
\alias{ATShtilde}
\title{ATS for vectors transformed with the function htilde}
\usage{
ATShtilde(N, X, C, r, a, d, p)
}
\arguments{
\item{N}{sample size}

\item{X}{matrix containing the bootstrap observations as columns}

\item{C}{the hypothesis matrix}

\item{r}{vectorised empirical covariance matrix of the original data}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}
}
\value{
a scalar, the value of the ATS
}
\description{
A function which calculates the Anova type statistic based on
the transformation htilde
}
