% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{ATSwS}
\alias{ATSwS}
\title{Weighted sum to calculate the asymptotic distribution of the ATS}
\usage{
ATSwS(A, repetitions)

ATSwS(A, repetitions)
}
\arguments{
\item{A}{the matrix which is used to calculate the ATS}

\item{repetitions}{the number of realisations of the weighted sum}
}
\value{
a vector containing the values of the weighted sum for a number of
repetitions

a vector containing the values of the weighted sum for a number of
repetitions
}
\description{
Generating a number of realisations of weighted sums of chi-square
distributed random variables which is the asymptotic distribution of the ATS

Generating a number of realisations of weighted sums of chi-square
distributed random variables which is the asymptotic distribution of the ATS
}
