% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrap for one group}
\usage{
Bootstrap(N.sim, n1, C, MSrootHatCov)

Bootstrap(N.sim, n1, C, MSrootHatCov)
}
\arguments{
\item{N.sim}{control variable for using sapply}

\item{n1}{a scalar, declaring the sample size for the bootstrap sample}

\item{C}{a hypothesis matrix for calculating the ATS}

\item{MSrootHatCov}{matrix root of the covariance matrix HatCov, to generate
the bootstrap sample}
}
\value{
a scalar, the value of the ATS
}
\description{
This function generates n1 normal distributed random vectors
with covariance matrix HatCov, which matrix root MSrootHatCov is given. For
the generated bootstrap sample the value of the ATS is calculated
}
