% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCovarianceMatrices.R
\name{Bootstrapg}
\alias{Bootstrapg}
\title{Bootstrap using the transformation g for one group}
\usage{
Bootstrapg(N.sim, n1, a, d, p, C, MSrootHatCov, vX)
}
\arguments{
\item{N.sim}{control variable for using sapply}

\item{n1}{a scalar, declaring the sample size for the bootstrap sample}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}

\item{C}{a hypothesis matrix for calculating the ATS}

\item{MSrootHatCov}{matrix root of the covariance matrix HatCov, to generate
the bootstrap sample}

\item{vX}{the expectation vector for the bootstrap sample}
}
\value{
a scalar, the value of the ATS
}
\description{
This function generates n1 normal distributed random vectors
with covariance matrix HatCov, which matrix root MSrootHatCov is given and
expectation vector vX. For the generated bootstrap sample the value of the
ATS based on transformation g is calculated
}
