% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{Bootstraphtilde}
\alias{Bootstraphtilde}
\title{Bootstrap using the transformation htilde for one group}
\usage{
Bootstraphtilde(N.sim, n1, a, d, p, C, MSUpsidv, vCorData)
}
\arguments{
\item{N.sim}{control variable for using sapply}

\item{n1}{a scalar, declaring the sample size for the bootstrap sample}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}

\item{C}{a hypothesis matrix for calculating the ATS}

\item{MSUpsidv}{matrix root of the covariance matrix Upsidv, to generate
the bootstrap sample}

\item{vCorData}{the expectation vector for the bootstrap sample}
}
\value{
a scalar, the value of the ATS
}
\description{
This function generates n1 normal distributed random vectors
with covariance matrix Upsidv, which matrix root is given and expectation
vector vCorData. For the generated bootstrap sample the value of the
ATS based on transformation h is calculated.
}
