% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{dvech}
\alias{dvech}
\title{Diagonal vectorisation including the diagonal}
\usage{
dvech(X, a, d, p)
}
\arguments{
\item{X}{quadratic matrix which should be vectorized}

\item{a}{vector containing the indices which belong to the diagonal of the
matrix}

\item{d}{dimension of the matrix which should be vectorised}

\item{p}{dimension of the vectorised matrix}
}
\description{
Diagonal vectorisation of the upper triangular part of a matrix
containing the diagonal elements.
}
