% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{dvechp}
\alias{dvechp}
\title{Diagonal vectorisation excluding the diagonal}
\usage{
dvechp(X, a, d, pu)
}
\arguments{
\item{X}{quadratic matrix which should be diagonalized}

\item{a}{vector containing the indices which belong to the diagonal of the
matrix}

\item{d}{dimension of the matrix which should be vectorised}

\item{pu}{dimension of the vectorised matrix}
}
\description{
Diagonal vectorisation of the upper triangular part of a matrix
without the diagonal elements.
}
