% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCovarianceMatrices.R
\name{g}
\alias{g}
\title{Transformation of the vectorised covariance matrix by quotients of means}
\usage{
g(v, a, d)
}
\arguments{
\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{x}{vectorised covariance matrix which should be transformed}
}
\value{
a transformed vector
}
\description{
A function which calculates the mean of the secondary diagonals
and divide them through the next one.Since the elements can be negative, for
the denominator absolute valued are used.
}
