% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{gData}
\alias{gData}
\title{Function to generate normal distributed bootstrap observations
A function to generate nv normal distributed random vectors with expectation
zero and covariance matrix Sigma, while the matrix root of Sigma is given.}
\usage{
gData(WSigma, nv)

gData(WSigma, nv)
}
\description{
Function to generate normal distributed bootstrap observations
A function to generate nv normal distributed random vectors with expectation
zero and covariance matrix Sigma, while the matrix root of Sigma is given.

Function to generate normal distributed bootstrap observations
A function to generate nv normal distributed random vectors with expectation
zero and covariance matrix Sigma, while the matrix root of Sigma is given.
}
