% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{htilde}
\alias{htilde}
\title{Root transformation of the vectorised correlation matrix}
\usage{
htilde(x, a, d)
}
\arguments{
\item{x}{diagonal vectorised correlation matrix which should be transformed}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}
}
\value{
a transformed vector
}
\description{
A function which calculates the roots of a vectorised covariance matrix, while
the roots increasing, so square root for the first secondary diagonals, third
root for the second secondary diagonal and so on. For roots with even order the
absolute value of the argument is used, since the arguments can be negative.
}
