% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCovarianceMatrices.R
\name{Jacobiang}
\alias{Jacobiang}
\title{Jacobian matrix for the function g}
\usage{
Jacobiang(X, a, d, p)
}
\arguments{
\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}

\item{x}{vectorised covariance matrix for which the Jacobian matrix is applied}
}
\value{
the Jacobian matrix applied for the given vector
}
\description{
A function which calculates the Jacobian matrix for
transformation function based on quotients of means. This Jacobian matrix is
applied for a given vector
}
