% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCovarianceMatrices.R
\name{Jacobianh}
\alias{Jacobianh}
\title{Jacobian matrix for the function h}
\usage{
Jacobianh(x, a, d, p)
}
\arguments{
\item{x}{vectorised covariance matrix  for which the Jacobian matrix is applied}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}
}
\value{
the Jacobian matrix applied for the given vector
}
\description{
A function which calculates the Jacobian matrix for the root
transformation h  applied for a given vector
}
