% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{Jacobianhtilde}
\alias{Jacobianhtilde}
\title{Jacobian matrix for the function htilde}
\usage{
Jacobianhtilde(x, a, d, p)
}
\arguments{
\item{x}{vectorised covariance matrix for which the Jacobian matrix is applied}

\item{a}{vector containing the indices which belong to the diagonal of the
covariance matrix}

\item{d}{dimension of the covariance matrix}

\item{p}{dimension of the vectorised matrix}
}
\value{
the Jacobian matrix applied for the given vector
}
\description{
A function which calculates the Jacobian matrix for the root
transformation htilde applied for a given vector
}
