% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{MSroot}
\alias{MSroot}
\title{Square root of a matrix}
\usage{
MSroot(X)

MSroot(X)
}
\description{
Square root of a matrix

Square root of a matrix
}
