% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{Pd}
\alias{Pd}
\title{Centering matrix}
\usage{
Pd(d)

Pd(d)
}
\arguments{
\item{d}{a scalar, characterizing the matrix and set its dimension}
}
\value{
the centering matrix

the centering matrix
}
\description{
matrix Pd for testing equality of the d components of a vector

matrix Pd for testing equality of the d components of a vector
}
