% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{QF}
\alias{QF}
\title{Quadratic form for vectors and matrices}
\usage{
QF(A, B)
}
\description{
Quadratic form for vectors and matrices
}
