% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{Qvech}
\alias{Qvech}
\title{Auxiliary function to calculate the covariance of the vectorized correlation matrix}
\usage{
Qvech(X, n)

Qvech(X, n)
}
\description{
Auxiliary function to calculate the covariance of the vectorized correlation matrix

Auxiliary function to calculate the covariance of the vectorized correlation matrix
}
