% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{Tayapp}
\alias{Tayapp}
\title{Function for the Taylor-based Monte-Carlo-approximation for one group}
\usage{
Tayapp(
  repetitions,
  C,
  MSrootStUpsi,
  CorData,
  MvrH1,
  Trace,
  M,
  L,
  P,
  Q,
  Atilde,
  n1
)
}
\arguments{
\item{repetitions}{a number specifying the number of runs for the approximation}

\item{C}{the used hypothesis matrix}

\item{MSrootStUpsi}{the matrix root of the covariance matrix for the Taylor
observations}

\item{MvrH1}{an auxiliary matrix for the transformation from vectorised covariances
to vectorized correlations}

\item{Trace}{a trace used in the ATS for the test statistic}

\item{M}{an auxiliary matrix for the transformation from vectorised covariances
to vectorized correlations}

\item{L}{an auxiliary matrix for the transformation from vectorised covariances
to vectorized correlations}

\item{P}{an auxiliary matrix for the transformation from vectorised covariances
to vectorized correlations}

\item{Q}{an auxiliary matrix for the transformation from vectorised covariances
to vectorized correlations}

\item{Atilde}{an auxiliary matrix for the transformation from row-wise vectorisation
to diagonalwise vectorisation}

\item{n1}{the sample size, a scalar}

\item{p}{the dimension of the vectorized matrix}

\item{Cordata}{the calculated correlation matrix}
}
\value{
a matrix containing the values of the Taylor ATS for a number of repetitions
}
\description{
An auxiliary function to calculate the values for the
Taylor-based Monte-Carlo-approximation for one group.  After receiving some
auxiliary matrices and data, the Monte-Carlo observations are generated and
different parts of the final sum are defined. Based on this a number of the
Taylor-based ATS are calculated, where the number can be chosen.
}
