% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{TestCorStructure}
\alias{TestCorStructure}
\title{Test for the correlation matrix of data regarding their structure}
\usage{
TestCorStructure(X, structure, method = "BT", repetitions = 1000, seed = 0)
}
\arguments{
\item{X}{a matrix containing the observation vectors as columns}

\item{structure}{a character specifying the structure regarding them the
covariance matrix should be checked. Options are "Hautoregressive",
"diagonal", "Hcompoundsymmetry" and "Htoeplitz", where H stands for Heterogenous.}

\item{method}{a character, to chose whether bootstrap("BT"),
Monte-Carlo-technique("MC") or a Taylor-based Monte-Carlos-approach("Tay")
is used, while bootstrap is the predefined method.}

\item{repetitions}{a scalar, indicate the number of runs for the chosen method.
The predefined value is 1,000, and the number should not be below 500.}

\item{seed}{A seed, if it should be set for reproducibility. Predefined values
is 0, which means no seed is set. A chosen seed is deleted at the end.}
}
\value{
a list containing the p-value, the value of the test statistic and the
value of the estimated covariance matrix used in the test
}
\description{
With this function the covariance matrix of data can be checked
for one of the usual structures. Depending on the chosen method a bootstrap or
Monte-Carlo-technique is used to calculate p-value of the ATS based on a
specified number of runs.
}
\examples{
X=matrix(rnorm(5*20),5,20)
TestCorStructure(X,structure="Htoeplitz",method="MC")
}
