% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{tvar}
\alias{tvar}
\title{Function to calculate the variance of transposed observations}
\usage{
tvar(X)

tvar(X)
}
\description{
Function to calculate the variance of transposed observations

Function to calculate the variance of transposed observations
}
