% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R
\name{vechp}
\alias{vechp}
\title{Upper triangular vectorisation without the diagonal}
\usage{
vechp(x)
}
\description{
Row-wise vectorisation of the upper triangular part of a matrix without the
diagonal elements.
}
