% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingStructuresofCorrelationMatrices.R,
%   R/TestingStructuresofCovarianceMatrices.R
\name{vtcrossprod}
\alias{vtcrossprod}
\title{Function to calculate vech(X t(X))}
\usage{
vtcrossprod(X)

vtcrossprod(X)
}
\description{
Function to calculate vech(X t(X))

Function to calculate vech(X t(X))
}
